Big refactor, use real c++ types

+++ MetalOS.lib/PageTables.cpp
@@ -1,5 +1,4 @@
 #include "PageTables.h"
-#include "Kernel.h"
 
 #define PhysicalToVirtual(x) ((x) + m_virtualOffset)
 
@@ -16,53 +15,53 @@
 
 //extern EFI_STATUS Keywait(const CHAR16* String);
 
-PageTables::PageTables(UINT64 physicalAddress) : m_physicalAddress(physicalAddress), m_pool(nullptr), m_virtualOffset(0)
+PageTables::PageTables(uintptr_t physicalAddress) : m_physicalAddress(physicalAddress), m_pool(nullptr), m_virtualOffset(0)
 {
 	
 }
 
-bool PageTables::MapUserPages(UINT64 virtualAddress, UINT64 physicalAddress, UINT32 count)
+bool PageTables::MapUserPages(uintptr_t virtualAddress, uintptr_t physicalAddress, uint32_t count)
 {
 	return MapPage(virtualAddress, physicalAddress, count, false);
 }
 
-bool PageTables::MapKernelPages(UINT64 virtualAddress, UINT64 physicalAddress, UINT32 count)
+bool PageTables::MapKernelPages(uintptr_t virtualAddress, uintptr_t physicalAddress, uint32_t count)
 {
 	return MapPage(virtualAddress, physicalAddress, count, true);
 }
 
-void PageTables::SetVirtualOffset(UINT64 virtualOffset)
+void PageTables::SetVirtualOffset(uintptr_t virtualOffset)
 {
 	m_virtualOffset = virtualOffset;
 }
 
 //This should be rewritten to map blocks instead of pages like this, but TODO etc
-bool PageTables::MapPage(UINT64 virtualAddress, UINT64 physicalAddress, UINT32 count, bool global)
+bool PageTables::MapPage(uintptr_t virtualAddress, uintptr_t physicalAddress, uint32_t count, bool global)
 {
 	Print(L"V: 0x%q P: 0x%q C: 0x%q G: %d VO: 0x%q\r\n", virtualAddress, physicalAddress, count, global, m_virtualOffset);
 	
 	//TODO: error code?
 	//Assert(m_pool != nullptr);
 
-	for (UINT32 i = 0; i < count; i++)
+	for (size_t i = 0; i < count; i++)
 	{
-		if (!MapPage(virtualAddress + ((UINT64)i << PAGE_SHIFT), physicalAddress + ((UINT64)i << PAGE_SHIFT), global))
+		if (!MapPage(virtualAddress + ((uintptr_t)i << PAGE_SHIFT), physicalAddress + ((uintptr_t)i << PAGE_SHIFT), global))
 			return false;
 	}
 
 	return true;
 }
 
-bool PageTables::MapPage(UINT64 virtualAddress, UINT64 physicalAddress, bool global)
+bool PageTables::MapPage(uintptr_t virtualAddress, uintptr_t physicalAddress, bool global)
 {
-	UINT64 newPageAddress;
-	UINT32 offset = virtualAddress & 0xFFF;
+	uintptr_t newPageAddress;
+	uint32_t offset = virtualAddress & 0xFFF;
 
 	//TODO: error code?
 	//Assert(m_pool != nullptr);
 
 	PPML4E l4 = (PPML4E)PhysicalToVirtual(m_physicalAddress);
-	UINT32 l4Index = (virtualAddress >> 39) & 0x1FF;
+	uint32_t l4Index = (virtualAddress >> 39) & 0x1FF;
 	Print(L"L4: 0x%q, Index: %d\r\n", l4, l4Index);
 
 	if (l4[l4Index].Value == 0)
@@ -80,7 +79,7 @@ bool PageTables::MapPage(UINT64 virtualAddress, UINT64 physicalAddress, bool glo
 		Print(L"  0x%q:", l4[l4Index].Value);
 	}
 	PPDPTE l3 = (PPDPTE)(PhysicalToVirtual(l4[l4Index].Value & ~0xFFF));
-	UINT32 l3Index = (virtualAddress >> 30) & 0x1FF;
+	uint32_t l3Index = (virtualAddress >> 30) & 0x1FF;
 	Print(L"L3: 0x%q, Index: %d\r\n", l3, l3Index);
 	Keywait(L"asd2\r\n");
 
@@ -99,7 +98,7 @@ bool PageTables::MapPage(UINT64 virtualAddress, UINT64 physicalAddress, bool glo
 	}
 	PPDE l2 = (PPDE)(PhysicalToVirtual(l3[l3Index].Value & ~0xFFF));
 	Keywait(L"asd\r\n");
-	UINT32 l2Index = (virtualAddress >> 21) & 0x1FF;
+	uint32_t l2Index = (virtualAddress >> 21) & 0x1FF;
 	if (l2[l2Index].Value == 0)
 	{
 		if (!m_pool->AllocatePage(&newPageAddress))
@@ -113,7 +112,7 @@ bool PageTables::MapPage(UINT64 virtualAddress, UINT64 physicalAddress, bool glo
 	}
 	PPTE l1 = (PPTE)(PhysicalToVirtual(l2[l2Index].Value & ~0xFFF));
 
-	UINT32 l1Index = (virtualAddress >> 12) & 0x1FF;
+	uint32_t l1Index = (virtualAddress >> 12) & 0x1FF;
 	//Assert(l1[l1Index].Value == 0);
 
 	//Make the connection
@@ -124,14 +123,14 @@ bool PageTables::MapPage(UINT64 virtualAddress, UINT64 physicalAddress, bool glo
 	l1[l1Index].Global = true;
 }
 
-UINT64 PageTables::ResolveAddress(UINT64 virtualAddress)
+uintptr_t PageTables::ResolveAddress(uintptr_t virtualAddress)
 {
 	//loading->WriteLineFormat("Resolving: 0x%16x", virtualAddress);
-	UINT32 offset = virtualAddress & 0xFFF;
-	UINT32 l1Index = (virtualAddress >> 12) & 0x1FF;
-	UINT32 l2Index = (virtualAddress >> 21) & 0x1FF;
-	UINT32 l3Index = (virtualAddress >> 30) & 0x1FF;
-	UINT32 l4Index = (virtualAddress >> 39) & 0x1FF;
+	uint32_t offset = virtualAddress & 0xFFF;
+	uint32_t l1Index = (virtualAddress >> 12) & 0x1FF;
+	uint32_t l2Index = (virtualAddress >> 21) & 0x1FF;
+	uint32_t l3Index = (virtualAddress >> 30) & 0x1FF;
+	uint32_t l4Index = (virtualAddress >> 39) & 0x1FF;
 
 	PPML4E l4 = (PPML4E)PhysicalToVirtual(m_physicalAddress);
 	Print(L"L4: 0x%q, P: %d, RW: %d S: %d -\r\n", l4[l4Index].Value, l4[l4Index].Present, l4[l4Index].ReadWrite, l4[l4Index].UserSupervisor);
@@ -153,13 +152,13 @@ UINT64 PageTables::ResolveAddress(UINT64 virtualAddress)
 	return (l1[l1Index].Value & ~0xFFF) + offset;
 }
 
-bool PageTables::EnableWrite(UINT64 virtualAddress)
+bool PageTables::EnableWrite(uintptr_t virtualAddress)
 {
-	UINT32 offset = virtualAddress & 0xFFF;
-	UINT32 l1Index = (virtualAddress >> 12) & 0x1FF;
-	UINT32 l2Index = (virtualAddress >> 21) & 0x1FF;
-	UINT32 l3Index = (virtualAddress >> 30) & 0x1FF;
-	UINT32 l4Index = (virtualAddress >> 39) & 0x1FF;
+	uint32_t offset = virtualAddress & 0xFFF;
+	uint32_t l1Index = (virtualAddress >> 12) & 0x1FF;
+	uint32_t l2Index = (virtualAddress >> 21) & 0x1FF;
+	uint32_t l3Index = (virtualAddress >> 30) & 0x1FF;
+	uint32_t l4Index = (virtualAddress >> 39) & 0x1FF;
 
 	PPML4E l4 = (PPML4E)PhysicalToVirtual(m_physicalAddress);
 	PPDPTE l3 = (PPDPTE)PhysicalToVirtual(l4[l4Index].Value & ~0xFFF);
